ai_backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
psycopg2==2.9.10
sqlalchemy==2.0.39